# src/swagger/swagger-output.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Item Simulator REST APIs",
    "description": "Chapter 3, 개인과제 - 아이템 시뮬레이터 "
  },
  "servers": [
    {
      "url": "/api/",
      "description": ""
    }
  ],
...
  "paths": {
...
    "/work/{characterId}": {
      "patch": {
        "tags": [
          "Characters Actions: 캐릭터행동관련"
        ],
        "summary": "게임머니 100 증가 API",
        "description": "[인증] 특정 캐릭터의 게임머니 100을 증가시킨다",
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer Token": []
          }
        ]
      }
    },
    "/buy/{characterId}/{itemCode}": {
      "post": {
        "tags": [
          "Characters Actions: 캐릭터행동관련"
        ],
        "summary": "아이템 구입 API",
        "description": "[인증] 아이템을 구입하여 캐릭터 인벤토리에 추가한다",
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer Token": []
          }
        ]
      }
    },
    "/sell/{characterId}/{invenId}": {
      "post": {
        "tags": [
          "Characters Actions: 캐릭터행동관련"
        ],
        "summary": "아이템 판매 API",
        "description": "[인증] 캐릭터 인벤토리의 아이템을 판매한다. (판매대금 60%)",
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer Token": []
          }
        ]
      }
    },
    "/{characterId}/inventory": {
      "get": {
        "tags": [
          "Characters Actions: 캐릭터행동관련"
        ],
        "summary": "인벤토리 목록 조회 API",
        "description": "[인증] 캐릭터 인벤토리 목록을 조회한다",
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer Token": []
          }
        ]
      }
    },
    "/dress/{characterId}/{invenId}": {
      "post": {
        "tags": [
          "Characters Actions: 캐릭터행동관련"
        ],
        "summary": "아이템 장착 API",
        "description": "[인증] 캐릭터 인벤토리에 있는 아이템을 장착한다",
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "Bearer Token": []
          }
        ]
      }
    },
    "/undress/{characterId}/{invenId}": {
      "post": {
        "tags": [
          "Characters Actions: 캐릭터행동관련"
        ],
        "summary": "아이템 탈착 API",
        "description": "[인증] 캐릭터의 장비 아이템을 탈착한다",
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "invenId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "Bearer Token": []
          }
        ]
      }
    },
    "/dressed/{characterId}": {
      "get": {
        "tags": [
          "Characters Actions: 캐릭터행동관련"
        ],
        "summary": "캐릭터 장비 목록 조회 API",
        "description": "[인증] 캐릭터의 장비 아이템 목록을 조회한다",
        "parameters": [
          {
            "name": "characterId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "Bearer Token": []
          }
        ]
      }
    }
  },
...
}